less-than: check for error when number of left, string on right
TEST: verify that the correct "expected numeric expression" error is
produced when there is a number on the left side operand and a string is
on the right side of a comparison operator.

This test initially failed, though the error was produced, because the
column and length was incorrect due to the wrong column and length being
passed to the expected numeric expression error instance (it was
incorrectly the column and length of the left side operand).

To get the column and length of the right side operand, the compile
comparison operator function was changed to call the compile
sub-expression function instead of calling the sub-expression member
directly because the former also returns the column and length of the
sub-expression in addition to the data type.