less-than: check for error when string on left, number of right
TEST: verify that the correct "expected string expression" error is
produced when there is a string on the left side and a number is on the
right side of a comparison operator.

This test initially failed because an "expected numeric expression"
error was produced, because this error was the only one implemented.

The comparison operator codes select function was changed to throw a new
empty expected string operand error if the right side operand data type
is not a string when the left side is a string.

The compile comparison operator function was changed to catch this new
error and throw a new expected string expression error class (same as
the expected numeric expression error class except for the message of
the error.

OTHER: added coverage of the dynamic destructor of the new expected
string expression error class (as a test, though it doesn't actually
test anything).