relational: implemented recreation of operator with string operands
TESTS: verify that each of the relation operators with string operands
are recreated correctly.

Each test was made to pass one at a time with the less-than,
greater-than, less-than or equal, and greater-than or equal operators by
replacing each string-string code instance null function pointer with a
pointer to the recreate binary operator function.

Note: recreation of the equality operators was not implemented since
their parsing of string operands were not implemented yet.