relational: implemented execution of operator with string operands
TESTS: verify that each of the relation operators with string operands
are executed correctly.

A new string compare function pointer type alias was added for a
function taking two standard constant string pointers.  An execute
compare string-string template function was added to that pops the right
side string operand, compares it to the left side operand and sets the
top integer from the boolean result.  This template takes a comparison
function as an arguments.

Each test was made to pass one at a time with the less-than,
greater-than, less-than or equal, and greater-than or equal operators by
implementing a string comparison function and replacing the null
function pointer with a pointer to the execute compare string-string
function with the comparison function as its template argument.