equal: implemented parsing with two string operands
TEST: verify that a equal expression with two string operands are parsed
to the end of the expression.

This test initially failed because the equal (equality) operator
expected a numeric expression as an operand.

The test was made to pass by changing the compile equality function from
calling the generic compile operator function to calling the just
implemented compile comparison operator function.