equality: implemented execution of operator with string operands
TESTS: verify that each of the equality operators with string operands
are executed correctly.

Each test was made to pass one at a time with the equal and not-equal
operators by implementing a string comparison function and replacing
the null function pointer with a pointer to the execute compare
string-string function with the comparison function as its template
argument.