cat: implemented parsing of concatenation operator
TEST: verify that a concatenation (add) expression with two string
operands is parsed to the end of the expression.

This test initially failed because the ADD operator only accepted
numeric operands.  The ADD operator is similar to the comparison
operators in taking string operands in addition to numeric operands.
However, they differ in what they return - comparison operators only
return an integer value, but ADD returns either a numeric or a string
value depending on the operands.

The hierarchy of the operator codes classes were rearranged.  A new
numeric-string operator codes class was added, which was just the
renaming of the comparison operator codes class.  The select function
was changed to set the return data types to that appropriate for the ADD
operator instead of a comparison operator (but for string operands, a
null data type is return since this functionality isn't implemented
yet).

A new comparison operator codes class was added with the numeric-string
operator codes class its base.  The select function is overridden, and
simply calls the number-string operator codes select function and sets
the return data type to integer (just like this class originally did
when it was based on the numeric operator codes class).

The add codes instance class was changed from the numeric operator codes
class to the new numeric-string operator codes class.  A string-string
add code instance was added (with null function pointers).  The compile
summation function was changed to call the compile comparison operator
function instead of the generic compile operator function.  The compile
summation function also handles the subtract operator, but this change
did not affect any of the 'current' subtract tests.

Since no longer handling just comparison operators, the compile
comparison operator function was renamed to the compile numeric-string
operator function.  To follow consistently, the generic compile operator
function was renamed to the compile numeric operator function.