sub: make sure subtract operator rejects string operands
TEST: verify that an "expected numeric expression" error is produced if
a SUB operator has string operands.

This test initially failed because the SUB operator, which is at the
same precedence level as the ADD operator so string data types could
now be passed to the select function.  The problem was that the numeric
operator codes class select function was loose in its selection of the
code because it assumed that it would only see either double or integer
data types.

The numeric operator codes select function was changed to check both the
left and right side data types before return the code info.  If the
correct data type is not found, then an empty expected numeric operand
error is thrown.

The SUB operator with string operands was now reporting an error, but
the error indicated that a numeric expression was expected on the right
side.  To correct this, the expected numeric operand error was renamed
to expected numeric right operand error, and a new expected numeric left
operand error was added.  The numeric operator codes select function was
changed to throw the expected numeric left operand error, and a catch
for this was added to compile numeric-string function, which then uses
the left side operand column and length to rethrow the expected numeric
expression error.