codes: refactored eliminating some code duplication
The code in the select function of the numeric operators codes and the
numeric data types section of the select function of the numeric-string
operators codes was essentially clones of each other.

To eliminate this duplication, a new numeric codes structure was added
for holding the four numeric codes.  If was given a constructor for
initializing the four member codes, and the four member codes in both
the numeric and numeric-string operator codes were replacing with an
instance of this, and their constructors changed to call the numeric
codes structure constructor.

A select function was added numeric codes structure for selecting the
appropriate code for the left and right side operand data types.  Both
the select functions in the  numeric and numeric-string operator codes
classes was changed to call this function.  Since both throw a different
error upon invalid data types, an error side argument was also added (an
enumeration with left and right enumerators).

OTHER: removed the unnecessary operator codes class name in front of
the info return structure on all uses.