cat: set to return new temporary string data type
TEST: verify that a concatenation expression returns the temporary
string data type.

There will be two types of strings, strings that come from constants or
variables including array elements that will have residence in the
associated dictionary, and temporary strings that will exist only during
the evaluation of a string expression.  At the end of a string
expression the temporary string will be assigned to a variable or array
element.

Codes that take string operands or arguments, will delete temporary
string no longer needed, but not regular strings, but will reuse the
temporary string is possible.

For the concatenation operator, one, both or none of the sides may be
a temporary string.  For the none case, a new temporary string will need
to be created.  For the both case, one of the temporary string can be
reused, but the other needs to be deleted.  For the one side cases, the
temporary string can be reused.

To get this test to pass, a new is temporary string function needed to
be added to the data type class along with a temporary string
enumerator, and a static creator function.  The numeric-string operator
codes class select function was modified to return a temporary string
return data type.

Another three codes will be needed for the add operator, but the above
changes were enough to get this test to pass.