expression: allow string string expressions within parentheses
TEST: verify that parentheses allow string expressions.

With only one string operator that returns a string (concatenation),
parentheses are not necessary to change the precedence of an operator.
However, parentheses are allow in string expression on GW-Basic and
QBasic, they will allowed here.

This test initially failed because the compile parentheses function was
only looking for a number when passed a null data type (when not called
from the compile function arguments function), and a null data type to
the compile expression function implies any number, but a number only,
hence the "expected numeric expression" error was produced.

To get this test to pass, a new compile function argument function was
added.  To keep the current tests passing, this function only needed to
get the open parenthesis character, skip white space, call the compile
expression with the expected data type passed from the compile function
arguments function, gets the closing parenthesis character, and returns
the data type from the compile expression function.

Since the expected data type argument of the compile parentheses
function was only used from the compile function arguments function, it
was removed, and the compile expression with no argument is called.
This allows string expressions within parentheses.  These changes caused
the test for no expression after an open parentheses to fail because
an "expected closing parentheses" error was produced instead of an
"expected numeric expression" error.  This was on a PRINT statement, so
the error should have been an "expected expression" since any type of
expression is supported by the PRINT command.  To correct this, a check
was added to the compile parentheses function to throw a new "expected
expression" error if a null data type is returned from the compile
expression.  A coverage test was added for this new error class.