expression: corrected error when missing numeric expression
TEST: verify an "expected numeric expression" error when there is a
missing expression in a numeric expression.

Unlike the previous test with a missing expression in a PRINT statement
that started with an open parentheses, if the expression started as a
numeric expression, an "expected expression" is not accurate.

To correct this problem, the compile parentheses needs to know what
type of expression preceded the open parentheses.  At the start of an
expression in a PRINT statement, any type could be expected.  But once a
numeric expression has started, then a numeric expression is expected
after an open parentheses.

One solution would be to simple add a flag.  This flag needs to have
three states (in others words, an enumeration), any expression, a
numeric expression, or a string expression.  This would required a
switch statement (and this is a code smell).

Instead of a enumeration, the abstract factory pattern was used.  Though
not technically abstract, the base expression error factory structure
contains a lone create error function, which simply throws an "expected
expression" error and takes the column as an argument.  The numeric
expression error factory derives from this base class and overrides the
create error function with one that throws an "expected number
expression" error.

The compile parentheses function was changed to call the create error
virtual function for the current expression error factory to throw the
error for the current expression type.

A unique pointer to an expression error factory was added to the
expression compiler implementation class, which is initialized to the
general expression error factory (for the default expression error).
The test contained the ADD operator, so after an ADD operator was
parsed, the expression error pointer is reset with a new instance to the
numeric expression error factory.