expression: corrected error when missing numeric expression (part 2)
TEST: verify an "expected numeric expression" error when there is a
missing expression in a pure numeric expression.

The previous test contained the ADD operator, which is handled by the
compile numeric-string operator function.  The pure numeric operators,
for example MULTIPLE, use the compile numeric operator function.  This
function was modified to reset the expression error factory to the
numeric expression error factory.