expression: corrected error when missing string expression
TEST: verify an "expected string expression" error when there is a
missing expression in a string expression.

The ADD operator can also be used with string operands, so if preceded
with a string, the error after an open parentheses with a missing
expression, an "expected numeric expression" is not correct.

A new string expression error factory was added where its create error
function throws an "expected string expression" error.  The compile
numeric-string operator function was changed to see if the left side
operand data type, and if numeric, then resets the expression error
factor to the numeric expression error factory, otherwise resets it to
the new string expression error factory.