expression: prevent wrong error with subtract operator and string
TEST: verify that the correct "expected numeric expression" error is
produced for the left operand of the SUB operator instead of an
"expected string expression" error after a parentheses with a missing
expression.

This test initially failed because the SUB operator is compiled in the
compile numeric-string operator function since it shares the same
precedence level as the ADD operator (which supports string operands).
The problem was that the missing expression was detected first since
the right side sub-expression operand is compiled before the operator
code is selected (the select function finds the error).

This function needed to check if the data type of the first operand is
valid for the operator before compiling the right side sub-expression.
This only applies when a SUB operator has a string for the left operand
as all the other operators (comparison and ADD) support strings (the
other operators are handled by the compile numeric operator function).

Another check first operand function could have been added to the codes
hierarchy, but that would have been a lot of changes for only one code
(SUB).  The select functions already had this functionality, except they
checked both operands.  The select function can be used if the left side
data type is passed for both left and right data types.  All operators
have a code where both operands are the same data type.

Therefore, a call to the select function was added right after getting
the codes with the left side data type passed for both arguments.  Any
return code is not needed as the select function is only being used for
its checking code (which throws an error when one is found).  This error
as caught and rethrown as an expected numeric expression error.

With these changes, it was no longer necessary to distinguish between
left and right numeric operand errors.  This was discovered when
checking the code coverage results.  Therefore the expected left and
right numeric operand empty error structures were replaced with a single
expected numeric operand error structure.  The error side argument of
numeric codes select functions was removed along with its enumeration.

The compile numeric-string operator function was getting rather unruly,
so parts of its code was extracted into a number of sub-functions.