cat: added compiling of temporary strings on the right side
TEST: verify that a concatenation expression is compiled with the
temporary string on the right side of the operator.

To get this test to compile, the add string-temporary string code
instance was added, which required a new string-temporary string
enumerator in the operator type enumeration.

To get this test to pass, a number of changes were required.  A new
string-temporary string code reference was added to the number-string
operator codes and the constructor.  With this changes, string-temporary
string code instances were needed for all of the comparison operators
(all with null function pointers to start).  A check for the right side
data type was added for a temporary string, returning the
string-temporary string code.

Following the pattern with the numeric codes being put into their own
structure, the string codes were also put into their own structure.  At
the moment, this structure only has two codes, but the other two will be
added shortly.