cat: added compiling of temporary strings on the left side
TEST: verify that a concatenation expression is compiled with the
temporary string on the left side of the operator.

To get this test to compile, the add temporary string-string code
instance was added, which required a new temporary string-string
enumerator in the operator type enumeration.

To get this test to pass, a number of changes were required.  A new
temporary string-string code reference was added to the string codes
structure, its constructor, and the the number-string operator codes
constructor.  With these changes, temporary string-string code instances
were needed for all of the comparison operators (all with null function
pointers to start).  A check for the left side data type was added for a
temporary string, returning the temporary string-string code (the right
side data type is not checked so an exception is not thrown for the left
operand check).