cat: added compiling of two temporary string operands
TEST: verify that a concatenation expression is compiled with two
temporary strings.

To get this test to compile, the add temporary string-temporary string
code instance was added, which required a new temporary string-temporary
string enumerator in the operator type enumeration.

To get this test to pass, a number of changes were required.  A new
temporary string-temporary string code reference was added to the string
codes structure, its constructor, and the the number-string operator
codes constructor.  With these changes, temporary string-temporary
string code instances were needed for all of the comparison operators
(all with null function pointers to start).  A check for the right side
data type was added for a string, returning the temporary string-string
code if a string, else a temporary string-temporary string code.