cat: added right side not a string check if left side temporary string
TEST: verify that an "expected string expression" error is produced if
the left side is a temporary string and the right side is not a string.

This test initially failed because the string codes select function was
accepting any data type other than a string on the right side if the
left side was a temporary string.

This test was made to pass by adding a check for a temporary string
after the check for a string before returning the temporary
string-temporary string code.

OTHER: renamed and re-tagged the current test case with 'compile' since
this test is large enough to stand on its own, and removed extra closing
parenthesis characters on two earlier tests (for these tests if didn't
matter if parsing terminated at the end of the stream or on this
character).