recreate: implemented recreation of operators with temporary strings
TEST: verify that concatenation expressions are recreated with string
and temporary string operands, and the comparison operators with
temporary string operands.

Each of these tests initially failed because all of these operators had
null pointers for the recreate function pointers.  Each test was added
one at a time and made to pass by replacing the null pointer of the
associated operator with the recreate binary operator function.

However, the first with a temporary string left side and string right
side continued to fail.  This was due to these codes not being added to
the precedence tables.  The test was made to pass be adding the code
value for the temporary string-string code to the vector of code values
in the number-string operator codes class.  Similarly, the first test
with a temporary string-temporary string failed until the code values
for these codes were added to the code values vector.