print: added compilation and recreation of temporary strings
TESTS: verify that a PRINT statement with a temporary string is
compiled and recreated correctly.

To get the first test (compiling) to compile, a new print temporary
string code instance was added with null function pointers.  To get this
test to pass, the compile print function was changed to check if the
data type is a string before adding a print string code, otherwise a
print temporary string code is added.

This test now passed, but now the temporary string recreation tests were
not failing.  These were previously passing because a print string code
was being put into the program, which are recreated to nothing.  to get
these tests to pass again, a pointer to the recreate nothing function
was given to the print temporary string code replacing the null pointer.

A separate recreation test was not necessary since this was covered in
the strings unit tests.  There are no implemented codes that put a
temporary string on the stack, so no test can be added to test the
print temporary string code execution yet.