cat: implemented execution of concatenation operator with  two strings
TEST: verify that a concatenation operator with two strings is executed
correctly.

The execute concatenation string-string function was implemented.  The
right side string operand is popped from the stack first.  A new
temporary result string instance is constructed with the value of the
left side string operand.  The right side string is appended to the
temporary result, and the top of the execution stack is set to the
result string.  This required a set top function with a standard string
pointer argument to be added to the executer class.  A pointer to this
function replaced the null function pointer in the add string string
code instance.

This test still crashed because there was no execute print temporary
string function.  One was implemented, which performs the same action
of the execute print string function with the addition of deleting the
temporary string instance that was popped from the stack.
The duplication between these two execute print string functions were
eliminated by extracting a new print top string function.