cat: implemented execution of remaining temporary string codes
TESTS: verify that concatenation operator with temporary strings as one
or both operands is executed correctly.

The three tests were added and made to pass one at a time.

The execute concatenate temporary string-string function reuses the left
side temporary string for the result.  The right side string operand is
popped from the stack, and appended directly to the temporary string on
top of the stack.  A new top temporary string function was needed to
return the string pointer on top of the stack so that it can be modified
since the top string function returns a constant string pointer
preventing modification.

The execute concatenate string-temporary string function reuses the
right side temporary string for the result.  A copy of the left side
string operand is made into a local string, and then swapped with the
result with the right side putting the left side string into the result.
The local string now with the right side string is then appended to the
result.  The top of the stack is then set to the result string pointer.

The execute concatenate temporary string-temporary string function
reused the left side temporary string for the result.  The right side
temporary string  is popped from the stack, and appended directly to the
temporary string on top of the stack.  The right side temporary string
is then deleted.

Units were run with valgrind to verify that there are no member leaks
associated with the temporary strings.