executer: changed stack handling of temporary strings
The handling of creating temporary strings was moved from the users
(execute concatenation and print item functions) into the executer class
member functions.  This includes the creation (allocation) and
destruction (deallocation), though the deallocation is handled
automatically because the temporary string pointer is put into a unique
pointer.  These changes simplify the code in the callers a bit.

The pointer to a string string value member of the executer stack item
was split into a constant string pointer named string value and a string
pointer named temporary string value.  The former is to represent
strings that should not be modified (for example constants), and can be
modified (temporary strings).

The string value and temporary string members are really just aliases of
each other.  The stack item constructor for string values was changed to
take a constant string pointer, and this constructor can be used to
initialize either string value members.  The top temporary string
function was changed to return the temporary string value member since
the return type of this function is a non-constant string value.

A new move top temporary string function was added that creates a
standard unique pointer of a standard string (new temporary string
alias) of the top temporary string value and returns it.  This does two
things.  It moves ownership of the temporary string from the stack (the
pointer on the stack is set to null) to the caller.  When it goes out of
scope in the caller, the temporary string member is deallocated
automatically (no delete operator call is needed).

This is utilized in the execute concatenate temporary string-temporary
string function where one of the temporary strings needs to be deleted,
and the print temporary string item where the temporary string needs to
be deleted.  The common print top string function was removed since not
much was left duplicated between the two print item string functions
(both now use different functions to get the top string).

OTHER: because of problems that occurred during the implementation of
these changes, the string unit tests were separated a bit and some were
given additional tags to allow running specific tests; the add string
code instances were renamed concatenate ('cat' for short) to mirror
their execution function names, the add code instance definitions were
moved above the execute concatenate just after the execute add functions
to separate the different functionality, and the spacing was added to
the execute concatenate functions so that they look less jumbled.