less-than: implemented generic comparison execution of temporary strings
TESTS: verify that the less than operator is executed correctly with
temporary strings on the left side, right side and both sides.

To get these tests to pass, generic execute comparison functions were
implemented, and a pointer to each generic execute comparison function
with the less than function as the template argument replaced the null
function pointers in the less than code instances.

Note: The tests were put into a new test case for relational operators
instead of adding to the test case containing the concatenation tests,
which was renamed.